Modified index.html, style.css + Added search-button.svg
Extracted svg to a separate file in order to make color change with hover possible. Replaced with img tag.

Created the skeleton design for the table with alternating color for separation and hovering so the user can know where the table starts and where the table ends.

Changed transition time for search bar.