Create tableSubmission.js
Contains a submission function that grabs the first name, last name, email, and level the user input through the form. This data is made into a JS object and sent as the body of a POST request with the URL '/studentData'.

Event listener added for the submit button for the relevant form.

Line 18: Use of preventDefault() in this location prevents the page from reloading every time the form is submitted by preventing this cancelable action.