Rename from tableSubmission.js
Added to do list.
Added a function that populates student data to the table. Used loops to populate but will maybe find a better way.
makeEditable function makes a particular row of a table editable. Still need to implement code to send to server and populate new data.
makeUneditable (inverse concept) is a work in progress.